// src/selectors.d.ts
export const SELECTORS: {
    CHAT_LIST_CONTAINER: string;
    CHAT_ITEM: string;
    CHAT_LAST_ITEM: string;
    CHAT_AREA: string;
    CHAT_CONTENT: string;
    CHAT_ROW: string;
    LOAD_MORE_BUTTON: string;
    PROFILE_BUTTON: string;
    PROFILE_PAGE: string;
    PHONE_NUMBER: string;
    WHATSAPP_CONTACT: string;
  };